+++ extensions/CuteHMI/Examples/Modbus/Controllers.3/CHANGES.md
@@ -3,3 +3,4 @@
 ## Version 3
 
 - This version has switched from CuteHMI.Modbus.3 to CuteHMI.Modbus.4.
+- Initialized status has been added